communication between fronend and backend is done and communication between backend and MariaDB established succesfully

+++ Githubhacktues12please/Entities/Multiplier.cs
@@ -2,6 +2,6 @@
 {
     public class Multiplier
     {
-        public decimal multiplier { get; set; }
+        public float multiplier { get; set; }
     }
 }
